Add basic typing for react-graph-vis
